backend/pyproject.toml
[project]
name = "confessions-backend"
version = "0.1.0"
description = "FastAPI backend for confessions application"
requires-python = ">=3.11"
dependencies = [
    "annotated-types==0.6.0",
    "anyio==3.7.1",
    "asyncpg>=0.29.0",
    "certifi==2023.7.22",
    "charset-normalizer==3.3.0",
    "click==8.1.7",
    "colorama==0.4.6",
    "emoji==2.8.0",
    "fastapi==0.103.2",
    "h11==0.14.0",
    "idna==3.4",
    "numpy==1.24.4",
    "opencv-python-headless==4.8.1.78",
    "packaging==23.2",
    "pillow==10.1.0",
    "pydantic==2.4.2",
    "pydantic-core==2.10.1",
    "python-dotenv>=1.0.0",
    "pytesseract==0.3.10",
    "python-multipart==0.0.6",
    "requests==2.31.0",
    "sniffio==1.3.0",
    "starlette==0.27.0",
    "typing-extensions==4.8.0",
    "urllib3==2.0.6",
    "uvicorn==0.23.2",
]
